faction_reason_exists_trigger
Implemented faction_reason_exists_trigger.
The political factions have a reason to exist if the Scourge, Burning Legion or Black Empire exists.
The Horde has a reason to exist if the Alliance exists.
The Alliance has a reason to exist if the Horde (as a title) is big or the Horde (as a faction) exists.