Fixed bug with Shath'Gral faiths not being allowed Ritual Cannibalism.
Commented out the line that blocked the religion from being able to use it for custom faiths. I did the same for any other lines with the comment "They have it by default" since I assumed that meant religions under that should be able to use the tenet.

I only commented them out since the issue was vague in what it wanted changed.
#114