Historical Councilors, Part 2
Cho'gall is a court chaplain of the Horde since year 583.
Kel'Thuzad is a spymaster of Dalaran since year 561.
Modera is a marshal of Dalaran since year 565.
Ansirem is a court chaplain of Dalaran since year 571.
Rhonin is a steward of Dalaran since year 584.
Drenden is a spymaster of Dalaran since year 563.
Kael'thas is a chancellor of Dalaran since year 583.