Plague Spreading Reworked
The Plague spreads to counties instead of provinces.
The Plague is limited to 30 counties.
Only humans are vulnerable to the Plague.
The Plague prefers counties with high Development and historical Plague regions.
The Plague Scheme doesn't work if the target has no vulnerable  counties.
Fixed the same county mentioned 3 times in the Plague target selecting event.
Localized plagued_county_modifier.