Blendshape + mesh fat and dwarf
The previous mesh had to be modify (vertex welded) and the previous version might not work as intended with blendshape and target meshes, I let it there marked as "previous" but i think the new one should be used.
I Also made a version with the blendshape already done, but let all the source mesh if need for you to do it yourself and fully control the name.