Fix/2 d/ai slop remover (#1793)
* AI Slop Remover

Removing AI art that I find, this is not done,

* Custom Ruler image

Custom Ruler image

* Changed too official art, changed names and removed art that is not used.

Removed some of the art that, as far as I can tell, have no use in the mod and changed some of the AI images to official art and sometimes art that I edited to fit their use. I also changed the names of some files to better fit a more general use or for future events to be added.

* Removed most if not all of the AI generated images

Removal of AI slop content