Perenold Tiara models
I made 3 different versions with 3 variants to each of the Perenold Tiara that I feel fit the design given. I'm not sure how I add the item to the items to the Barbershop so its all in the Debug Mode
Tilde -> Portrait Editor -> Under Ethnicity press randomize DNA to get a human test dummy -> Under Gene search for Headgear -> Go to the bottom of the Sub group and you will find Perenold Tiara 1-3, remember to adjust the strenght on it to see the variants to each Tiara design.

The variants are made to give some option of use but also to give others something to instruct the best posible model.
The model might look odd on a bald model but you can add hair under the Gene category to give it a try with it.
Things that need to be done:
Add the Tiara/s to the Barbershop
Add a use for the Tiara in game, whether its a item or just the crown attached to Alterac rule.